Presenting, slide and deck search
Presentations are generated server-side by slidewinder, then stored in a Presentations collection. This collection is routed to /collection/:_id on the server, where _id is a SHA1 hash of the generated HTML. This avoids identical dekcs being stored.

Slide and deck search are provided by Easy-Search - TODO is to connect them up to elastic.